Improve unit test consistency (#10)
Newer versions of pipenv drop the keyword argument 'project' in
`convert_deps_to_pip()`, which in itself was easy to fix, but the tests had
become a mess because of newly discovered security issues in the packages used
in the unit tests.

Fixed the call to `convert_deps_to_pip()`, spotted and fixed a bug in the
requirements StringIO part and added Python 3.12 testing to Tox. To avoid newly
discovered vulnerabilities to make tests fail, I used the `SafetyContext`
singleton to point SafetyDB to a local database for the unit tests.